Move background _utils.py to utils/_stats.py
